requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0
matplotlib>=3.7.0
PyQt6>=6.5.0
